+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## 1.1.0 - 2022-03-14
+### Added
+- Add the ability to pass custom options to the Sentry SDK by listening to the 
+  `SentryService::EVENT_DEFINE_SENTRY_SDK_CONFIGURATION` event
+- Honor the Craft [`httpProxy` general setting](https://craftcms.com/docs/3.x/config/config-settings.html#httpproxy)
+- Added console function to send a test exception
+
 ## 1.0.3 - 2020-10-29
 ### Fixed
 - Only set up Sentry after we've made sure we need to process the exception